pyproject.toml
[project]
name = "backtester"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "celery>=5.6.2",
    "fastapi>=0.129.0",
    "passlib>=1.7.4",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "redis>=7.2.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
]
